+++ packages/nimble-components/src/table/components/row/index.ts
@@ -3,11 +3,13 @@ import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
 import { styles } from './styles';
 import { template } from './template';
 import type {
+    TableActionMenuToggleEventDetail,
     TableCellState,
     TableDataRecord,
     TableFieldName
 } from '../../types';
 import type { TableColumn } from '../../../table-column/base';
+import type { MenuButtonToggleEventDetail } from '../../../menu-button/types';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -36,6 +38,12 @@ export class TableRow<
     @observable
     public columns: TableColumn[] = [];
 
+    @observable
+    public currentActionMenuColumn?: TableColumn;
+
+    @attr({ attribute: 'menu-open', mode: 'boolean' })
+    public menuOpen = false;
+
     @volatile
     public get columnStates(): ColumnState[] {
         return this.columns.map(column => {
@@ -62,6 +70,40 @@ export class TableRow<
 
             return { column, cellState };
         });
+    }
+
+    public onCellActionMenuBeforeToggle(
+        event: CustomEvent<MenuButtonToggleEventDetail>,
+        column: TableColumn
+    ): void {
+        this.currentActionMenuColumn = column;
+        this.emitToggleEvent(
+            'row-action-menu-beforetoggle',
+            event.detail,
+            column
+        );
+    }
+
+    public onCellActionMenuToggle(
+        event: CustomEvent<MenuButtonToggleEventDetail>,
+        column: TableColumn
+    ): void {
+        this.menuOpen = event.detail.newState;
+        this.emitToggleEvent('row-action-menu-toggle', event.detail, column);
+    }
+
+    private emitToggleEvent(
+        eventType: string,
+        menuButtonEventDetail: MenuButtonToggleEventDetail,
+        column: TableColumn
+    ): void {
+        const detail: TableActionMenuToggleEventDetail = {
+            newState: menuButtonEventDetail.newState,
+            oldState: menuButtonEventDetail.oldState,
+            recordIds: [this.recordId!],
+            columnId: column.columnId
+        };
+        this.$emit(eventType, detail);
     }
 
     private hasValidFieldNames(