+++ packages/nimble-components/src/table/template.ts
@@ -12,6 +12,7 @@ import type { Table } from '.';
 import { TableHeader } from './components/header';
 import { TableRow } from './components/row';
 import type { TableColumn } from '../table-column/base';
+import type { TableActionMenuToggleEventDetail } from './types';
 
 // prettier-ignore
 export const template = html<Table>`
@@ -29,7 +30,7 @@ export const template = html<Table>`
             </div>
             <div class="table-viewport" ${ref('viewport')}>
                 <div class="table-scroll" style="height: ${x => x.virtualizer.allRowsHeight}px;"></div>
-                <div class="table-row-container" role="rowgroup" style="transform: ${x => (x.virtualizer.rowContainerYOffset === 0 ? 'none' : `translateY(${x.virtualizer.rowContainerYOffset}px)`)};">
+                <div class="table-row-container" role="rowgroup" style="top: ${x => `${x.virtualizer.rowContainerYOffset}px;`}">
                     ${when(x => x.columns.length > 0 && x.canRenderRows, html<Table>`
                         ${repeat(x => x.virtualizer.visibleItems, html<VirtualItem, Table>`
                             <${DesignSystem.tagFor(TableRow)}
@@ -37,8 +38,18 @@ export const template = html<Table>`
                                 record-id="${(x, c) => c.parent.tableData[x.index]?.id}"
                                 :dataRecord="${(x, c) => c.parent.tableData[x.index]?.record}"
                                 :columns="${(_, c) => c.parent.columns}"
+                                @row-action-menu-beforetoggle="${(_, c) => c.parent.onRowActionMenuBeforeToggle(c.event as CustomEvent<TableActionMenuToggleEventDetail>)}"
+                                @row-action-menu-toggle="${(_, c) => c.parent.onRowActionMenuToggle(c.event as CustomEvent<TableActionMenuToggleEventDetail>)}"
                                 style="height: ${x => x.size}px;"
                             >
+                            ${when((x, c) => (c.parent as Table).openActionMenuRecordId === (c.parent as Table).tableData[x.index]?.id, html<VirtualItem, Table>`
+                                ${repeat((_, c) => (c.parent as Table).actionMenuSlots, html<string, Table>`
+                                    <slot
+                                        name="${x => x}"
+                                        slot="${x => `row-action-menu-${x}`}">
+                                    </slot>
+                                `)}
+                            `)}                        
                             </${DesignSystem.tagFor(TableRow)}>
                         `)}
                     `)}